+++ assets/styles/components/_sidebar.scss
@@ -123,7 +123,7 @@
         padding: 0;
 
         a {
-          @include kbin-btn-link;
+          @include mbin-btn-link;
         }
 
         p:last-of-type {
