+++ assets/styles/layout/_typo.scss
@@ -57,5 +57,5 @@ h6 {
 }
 
 .btn-link {
-  @include kbin-btn-link;
+  @include mbin-btn-link;
 }
