+++ assets/styles/mixins/kbin.scss
@@ -1,4 +1,4 @@
-@mixin kbin-btn-link {
+@mixin mbin-btn-link {
   background: none;
   border: 0;
   display: inline;
