+++ config/mbin_routes/domain.yaml
