+++ config/mbin_routes/domain_api.yaml
