+++ config/mbin_routes/entry.yaml
