+++ config/mbin_routes/entry_api.yaml
