+++ config/mbin_routes/magazine_api.yaml
