+++ config/mbin_routes/notification_api.yaml
