+++ config/mbin_routes/page.yaml
