+++ config/mbin_routes/search.yaml
