+++ config/mbin_routes/user_api.yaml
