+++ config/mbin_serialization/domain.yaml
