+++ config/mbin_serialization/entry.yaml
