+++ config/mbin_serialization/entry_comment.yaml
