+++ config/mbin_serialization/image.yaml
