+++ config/mbin_serialization/magazine.yaml
