+++ config/mbin_serialization/user.yaml
